Return exit code from script

+++ test/runtests.sh
@@ -23,8 +23,8 @@ source $WORKSPACE/.venv/$ACTIVATE_DIR/activate
 fair registry install
 fair registry start
 fair init --ci
-fair pull --local $WORKSPACE/examples/fdp/SEIRSconfig.yaml
-fair run --dirty --local $WORKSPACE/examples/fdp/SEIRSconfig.yaml
+if ! fair pull --local $WORKSPACE/examples/fdp/SEIRSconfig.yaml; then exit 1; fi
+if ! fair run --dirty --local $WORKSPACE/examples/fdp/SEIRSconfig.yaml; then exit 1; fi
 
 TEST_SCRIPT="$(printf ' %q' "$@")"
 echo Test: "$TEST_SCRIPT"
@@ -32,7 +32,7 @@ ESCAPED_SCRIPT=$(printf '%s\n' "$TEST_SCRIPT" | sed -e 's/[\,&]/\\&/g')
 echo Escaped test: "$ESCAPED_SCRIPT"
 sed -e "s,\$TEST_SCRIPT,$ESCAPED_SCRIPT," $TEST_DIR/pre_config.yaml > $TEST_DIR/config.yaml
 cat $TEST_DIR/config.yaml
-fair run --dirty --local --debug $TEST_DIR/config.yaml
+if ! fair run --dirty --local --debug $TEST_DIR/config.yaml; then exit 1; fi
 rm -f $TEST_DIR/config.yaml
 
 deactivate
